requirements_docker.txt
accelerate==1.6.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.3.1
Automat==25.4.16
backoff==2.2.1
banks==2.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
blobfile==3.0.0
boto3==1.37.37
botocore==1.37.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
Crawl4AI==0.5.0.post8
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==8.0.1
einops==0.8.1
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.12
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
FlashRank==0.2.10
Flask==3.1.0
flatbuffers==25.2.10
free_proxy==1.1.3
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.0
griffe==1.7.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
html2text==2025.4.15
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
humanize==4.12.2
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-aws==0.2.21
langchain-community==0.3.21
langchain-core==0.3.54
langchain-huggingface==0.1.2
langchain-mistralai==0.2.10
langchain-ollama==0.3.2
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.32
litellm==1.66.3
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-index==0.12.31
llama-index-agent-openai==0.4.6
llama-index-callbacks-langfuse==0.3.0
llama-index-cli==0.4.1
llama-index-core==0.12.31
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.37
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-weaviate==1.3.1
llama-parse==0.6.12
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
minify_html==0.16.4
mmh3==5.1.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
mypy-extensions==1.0.0
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
ollama==0.4.8
onnxruntime==1.21.1
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==10.4.0
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
primp==0.15.0
propcache==0.3.1
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
pyiceberg==0.9.0
pymongo==4.12.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdf==5.4.0
pyperclip==1.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
pywin32==310
PyYAML==6.0.2
queuelib==1.8.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3fs==2025.3.2
s3transfer==0.11.5
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scrapegraph_py==1.12.0
scrapegraphai==1.48.0
Scrapy==2.12.0
scrapy-rotating-proxies==0.6.2
selenium==4.31.0
semchunk==3.2.1
sentence-transformers==4.1.0
service-identity==24.2.0
setuptools==78.1.0
simpleeval==1.0.3
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.2
strictyaml==1.7.3
striprtf==0.0.26
sympy==1.13.1
tavily-python==0.5.4
tenacity==9.1.2
text-unidecode==1.3
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
thrift==0.21.0
tiktoken==0.9.0
tldextract==5.2.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
validators==0.34.0
w3lib==2.3.1
weaviate-client==4.13.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
